slice functions moved to relevant files, and offset/limit query parameter functions moved to api.go
