+++ api/codes.go
@@ -6,7 +6,6 @@ import (
 	"sort"
 
 	"github.com/ONSdigital/dp-code-list-api/models"
-	"github.com/ONSdigital/dp-code-list-api/utils"
 	"github.com/ONSdigital/log.go/log"
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
@@ -22,7 +21,7 @@ func (c *CodeListAPI) getCodes(w http.ResponseWriter, r *http.Request) {
 	log.Event(ctx, "getCodes endpoint: attempting to get edition codes", log.INFO, data)
 
 	// get limit from query parameters, or default value
-	limit, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
+	limit, err := GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
 	if err != nil {
 		log.Event(ctx, "failed to obtain limit from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain limit", log.Data{"limit": limit}, err, w)
@@ -30,7 +29,7 @@ func (c *CodeListAPI) getCodes(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// get offset from query parameters, or default value
-	offset, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
+	offset, err := GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
 	if err != nil {
 		log.Event(ctx, "failed to obtain offset from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain offset", log.Data{"offset": offset}, err, w)
@@ -60,7 +59,7 @@ func (c *CodeListAPI) getCodes(w http.ResponseWriter, r *http.Request) {
 		return codes.Items[i].ID < codes.Items[j].ID
 	})
 
-	slicedResults := utils.CodesSlice(codes.Items, offset, limit)
+	slicedResults := codesSlice(codes.Items, offset, limit)
 
 	count := len(slicedResults)
 	codes.Count = count
@@ -119,4 +118,16 @@ func (c *CodeListAPI) getCode(w http.ResponseWriter, r *http.Request) {
 	}
 
 	log.Event(ctx, "getCode endpoint: request successful", log.INFO, data)
+}
+
+func codesSlice(full []models.Code, offset, limit int) (sliced []models.Code) {
+	end := offset + limit
+	if limit == 0 || end > len(full) {
+		end = len(full)
+	}
+
+	if offset > len(full) {
+		return []models.Code{}
+	}
+	return full[offset:end]
 }
