+++ api/datasets.go
@@ -6,7 +6,6 @@ import (
 	"sort"
 
 	"github.com/ONSdigital/dp-code-list-api/models"
-	"github.com/ONSdigital/dp-code-list-api/utils"
 	"github.com/ONSdigital/log.go/log"
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
@@ -23,7 +22,7 @@ func (c *CodeListAPI) getCodeDatasets(w http.ResponseWriter, r *http.Request) {
 	log.Event(ctx, "getCodeDatasets endpoint: attempting to find datasets related to code", log.INFO, logData)
 
 	// get limit from query parameters, or default value
-	limit, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
+	limit, err := GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
 	if err != nil {
 		log.Event(ctx, "failed to obtain limit from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain limit", log.Data{"limit": limit}, err, w)
@@ -31,7 +30,7 @@ func (c *CodeListAPI) getCodeDatasets(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// get offset from query parameters, or default value
-	offset, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
+	offset, err := GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
 	if err != nil {
 		log.Event(ctx, "failed to obtain offset from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain offset", log.Data{"offset": offset}, err, w)
@@ -58,7 +57,7 @@ func (c *CodeListAPI) getCodeDatasets(w http.ResponseWriter, r *http.Request) {
 		return datasets.Items[i].ID < datasets.Items[j].ID
 	})
 
-	slicedResults := utils.DatasetsSlice(datasets.Items, offset, limit)
+	slicedResults := datasetsSlice(datasets.Items, offset, limit)
 
 	count := len(slicedResults)
 	datasets.Count = count
@@ -78,4 +77,16 @@ func (c *CodeListAPI) getCodeDatasets(w http.ResponseWriter, r *http.Request) {
 	}
 
 	log.Event(ctx, "getCodeDatasets endpoint: request successful", log.INFO, logData)
+}
+
+func datasetsSlice(full []models.Dataset, offset, limit int) (sliced []models.Dataset) {
+	end := offset + limit
+	if limit == 0 || end > len(full) {
+		end = len(full)
+	}
+
+	if offset > len(full) {
+		return []models.Dataset{}
+	}
+	return full[offset:end]
 }
