+++ api/editions.go
@@ -6,7 +6,6 @@ import (
 	"sort"
 
 	"github.com/ONSdigital/dp-code-list-api/models"
-	"github.com/ONSdigital/dp-code-list-api/utils"
 	"github.com/ONSdigital/log.go/log"
 	"github.com/gorilla/mux"
 	"github.com/pkg/errors"
@@ -21,7 +20,7 @@ func (c *CodeListAPI) getEditions(w http.ResponseWriter, r *http.Request) {
 	log.Event(ctx, "getEditions endpoint: attempting to find editions", log.INFO, data)
 
 	// get limit from query parameters, or default value
-	limit, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
+	limit, err := GetPositiveIntQueryParameter(r.URL.Query(), "limit", c.defaultLimit)
 	if err != nil {
 		log.Event(ctx, "failed to obtain limit from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain limit", log.Data{"limit": limit}, err, w)
@@ -29,7 +28,7 @@ func (c *CodeListAPI) getEditions(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// get offset from query parameters, or default value
-	offset, err := utils.GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
+	offset, err := GetPositiveIntQueryParameter(r.URL.Query(), "offset", c.defaultOffset)
 	if err != nil {
 		log.Event(ctx, "failed to obtain offset from request query parameters", log.ERROR)
 		handleError(ctx, "failed to obtain offset", log.Data{"offset": offset}, err, w)
@@ -59,7 +58,7 @@ func (c *CodeListAPI) getEditions(w http.ResponseWriter, r *http.Request) {
 		return editions.Items[i].ID < editions.Items[j].ID
 	})
 
-	slicedResults := utils.EditionsSlice(editions.Items, offset, limit)
+	slicedResults := editionsSlice(editions.Items, offset, limit)
 
 	count := len(slicedResults)
 	editions.Count = count
@@ -113,4 +112,16 @@ func (c *CodeListAPI) getEdition(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 	log.Event(r.Context(), "retrieved codelist edition", log.INFO, log.Data{"code_list_id": id, edition: edition})
+}
+
+func editionsSlice(full []models.Edition, offset, limit int) (sliced []models.Edition) {
+	end := offset + limit
+	if limit == 0 || end > len(full) {
+		end = len(full)
+	}
+
+	if offset > len(full) {
+		return []models.Edition{}
+	}
+	return full[offset:end]
 }
