+++ go.mod
@@ -3,7 +3,6 @@ module github.com/ONSdigital/dp-code-list-api
 go 1.13
 
 require (
-	github.com/ONSdigital/dp-dataset-api v1.25.0
 	github.com/ONSdigital/dp-graph/v2 v2.6.1
 	github.com/ONSdigital/dp-healthcheck v1.0.5
 	github.com/ONSdigital/dp-net v1.0.10
